Mix with Brazilium and Voluptous
- Added CEP, CNPJ, CPF and email validator functions
- TODO add coverage to the rest
