+++ requirements.txt
@@ -4,3 +4,4 @@ voluptuous==0.8.11
 coveralls==1.1
 scrutinizer-ocular==1.0.3
 mock==2.0.0
+brazilnum==0.8.7
